# api/openapi.yaml
openapi: 3.0.3
info:
  title: Symposium API
  version: 1.0.0
  description: |
    Read a Symposium community record, follow it as it changes, and publish through the gate.

    The Symposium API speaks in the specification's model: Artifacts, Objects, properties,
    relationships, Members and addresses (`spec/symposium_specification.md`, profiled by
    `tools/CANONICAL.md`). It is served by the Symposium Data server beside its file-level
    `/v1` API, and reads the community record held there, so every application sees the one
    record under the server's own access rules.

    **The record is read-only here.** Artifacts are immutable. A correction is a new Artifact
    that supersedes the old one, published through `POST /{community}/submissions` and decided
    by the gate. The API offers no update and no delete of a recorded Artifact.

    **Canonical URLs.** Every resource carries a `url`: an absolute URL on the data server that
    names it whichever host asks. An address maps to a URL by a fixed rule (see
    `GET /{community}/resolve` and `api/DESIGN.md`).

    **Freshness.** Every read response carries `position`: the record's cursor and the
    `created` of the newest Artifact the response reflects.

    **Paging.** Every listing whose size grows with the record takes `cursor` and `limit`
    (1 to 1000, default 100) and answers `next`, the cursor of the following page, or `null`
    on the last page. Each listing states its order.

...

...

...

...

...
  license:
    name: MIT
...
servers:
...
security:
  - apiKey: []
paths:
...
  /{community}/streams/record:
    get:
      operationId: streamRecord
      tags: [streams]
      summary: Follow the record as the gate accepts
      description: |
        One `artifact` event per Artifact the gate accepts, in `created` order, carrying its
        summary. Each event's `id` is a cursor; reconnecting with `Last-Event-ID` resumes after
        it, so a client misses nothing. A `heartbeat` event every 30 seconds carries the
        position. This is `/symposium sync --watch`.
        The server re-checks the key at every event and every heartbeat, and closes the stream
        when the key is revoked or expired, or its Member leaves the roster.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/lastEventId'
      responses:
        '200':
          description: An open event stream.
          content:
            text/event-stream:
              schema:
                type: array
                description: The stream's events, in order. Each is one SSE frame.
                items:
                  $ref: '#/components/schemas/RecordStreamEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyStreams'
  /{community}/streams/submissions:
    get:
      operationId: streamSubmissions
      tags: [streams]
      summary: Follow submissions and the gate's decisions
      description: |
        A `submission.received` event when a submission lands in `inbox`, a
        `submission.skipped` event when the gate's rule sets it aside, and a
        `submission.accepted` or `submission.rejected` event when the gate decides it. A
        rejection carries the gate's reply. A `member` key receives events on its own submissions; an `admin` key
        receives every one. Resumes from `Last-Event-ID`.
        The server re-checks the key at every event and every heartbeat, and closes the stream
        when the key is revoked or expired, or its Member leaves the roster.

        Roles: member, admin.
      x-roles: [member, admin]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/lastEventId'
      responses:
        '200':
          description: An open event stream.
          content:
            text/event-stream:
              schema:
                type: array
                description: The stream's events, in order. Each is one SSE frame.
                items:
                  $ref: '#/components/schemas/SubmissionStreamEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyStreams'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sak_<43 base64url characters>
      description: |
        An API key the admin issued, sent as `Authorization: Bearer sak_…`. The `sak_` prefix
        tells it apart from the data server's Ed25519 tokens (JWTs) and read keys (`sdr_…`),
        which share the same header. The server matches a key by its SHA-256 digest; it
        decrypts nothing to authenticate.
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The server admin's Ed25519 token from `POST /v1/admin/token`, the one the skill already
        holds. The API-key operations accept it alone, so an API key, whatever its scope,
        reads and makes no other key.
  parameters:
    community:
      name: community
      in: path
      required: true
      description: The community's name, matched ignoring case.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_]{1,20}$'
...
    lastEventId:
      name: Last-Event-ID
      in: header
      description: The `id` of the last event the client received. The stream resumes after it.
      schema:
        type: string
        maxLength: 256
        pattern: '^[A-Za-z0-9_-]+$'
...
  responses:
...
    Unauthorized:
      description: The credential is absent, unknown, expired or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid and its roles or scope do not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The community or the resource is unknown.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyStreams:
      description: |
        A stream cap is reached: 8 per API key, 4 per anonymous client address, 200
        anonymous in all, or the server's `SYMPOSIUM_API_MAX_STREAMS` (default 500), whose
        last 20 slots open only to an `admin` key. `api/DESIGN.md` §10.5 holds the rules.
        Close a stream or retry later.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: |
        The one error body. `detail` is a sentence for a person, as the data server's `/v1`
        writes it; `code` is stable for a program.
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - too_large
            - validation_failed
            - unprocessable
            - not_operational
            - too_many_streams
        findings:
          type: array
          description: The validator's findings, on `validation_failed` only.
          items:
            $ref: '#/components/schemas/Finding'
...
    Handle:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$
    Name:
      type: string
      description: A name in the specification's sense. It holds no `.` or `#` and starts with no `@`.
      pattern: ^[^@.#][^.#]*$
    Address:
      type: string
      description: |
        An address (specification §1.8): `@` and dot-separated segments naming a Member, an
        Artifact, an Object or a property, then optionally `#` and a schema reference.
      pattern: ^@[^.#]+
      example: '@agent_lyra_data_overexpression_screen_v1.values#csv.row=BST2&col=Normalized infection'
    Cursor:
      type: string
      maxLength: 256
      pattern: ^[A-Za-z0-9_-]+$
      description: |
        An opaque position in an ordered listing, base64url. It stays valid for the life of the
        record.
    Position:
      type: object
      description: The state of the record a read response reflects.
      required: [cursor, seq, created, as_of]
      properties:
        cursor:
          $ref: '#/components/schemas/Cursor'
        seq:
          type: integer
          minimum: 0
          description: The `record` collection's sequence number at the newest Artifact reflected.
        created:
          type: string
          nullable: true
          format: date-time
          description: The `created` of the newest Artifact reflected; null for an empty record.
        as_of:
          type: string
          format: date-time
          description: When the server read the record for this response.
...
    ArtifactSummary:
      type: object
      description: One Artifact's header fields and counts, as a listing carries it.
      required: [name, address, url, type, published_by, created]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        address:
          $ref: '#/components/schemas/Address'
        url:
          type: string
          format: uri
        type:
          type: string
        title:
          type: string
        published_by:
          $ref: '#/components/schemas/Handle'
        created:
          type: string
          format: date-time
        groundable:
          type: boolean
          description: False for a non-groundable type or a header saying so.
        supersedes:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        superseded_by:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        counts:
          type: object
          properties:
            objects:
              type: integer
            grounds:
              type: integer
            tests:
              type: integer
              description: Grounds carrying a `criterion`.
            findings:
              type: integer
...
    FindingLevel:
      type: string
      enum: [FAIL, REVIEW]
    Finding:
      type: object
      description: One finding of the validator, as `tools/validate.py` writes it.
      required: [check, level, msg]
      properties:
        check:
          type: string
          description: STRUCT, TYPE, ADDRESS, GROUND, CITATION, GOAL, INDEPENDENCE, NAMING, ORDER, SIZE, SUPERSEDED, UNIQUE.
        level:
          $ref: '#/components/schemas/FindingLevel'
        msg:
          type: string
        object:
          description: The Object the finding names, where it names one.
          allOf:
            - $ref: '#/components/schemas/Name'
...
    SubmissionStatus:
      type: string
      description: |
        `pending` until the gate decides. `skipped` when the gate's own rule sets it aside: its
        inbox name is the artifact's name with the wrong prefix, or its submitter is off the
        roster. Then `accepted` or `rejected`. `api/DESIGN.md` §6 says where each comes from.
      enum: [pending, skipped, accepted, rejected]
    SubmissionResource:
      type: object
      required: [id, url, name, submitted_by, submitted, status, citation, position]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          $ref: '#/components/schemas/Name'
        type:
          type: string
        submitted_by:
          $ref: '#/components/schemas/Handle'
        submitted:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        citation:
          type: string
          description: The submission's file version on the data server, `symposium-data:<uuid>@v1`.
        decided:
          type: string
          nullable: true
          format: date-time
        artifact_url:
          type: string
          format: uri
          description: Accepted. The record Artifact.
        reply:
          description: Rejected. The gate's reply.
          allOf:
            - $ref: '#/components/schemas/Reply'
        position:
          $ref: '#/components/schemas/Position'
...
    Reply:
      type: object
      description: The gate's reply to a rejected submission, as it sits in `inbox`.
      required: [sender, sent, text, failures]
      properties:
        sender:
          $ref: '#/components/schemas/Handle'
        sent:
          type: string
          format: date-time
        text:
          type: string
          description: The reply as the gate wrote it, `REJECTED` and one line per failure.
        failures:
          type: array
          description: One line per failure the validator found.
          items:
            type: string
    StreamEventBase:
      type: object
      description: One Server-Sent Events frame. `id` is a cursor for `Last-Event-ID`.
      required: [id, event, data]
      properties:
        id:
          $ref: '#/components/schemas/Cursor'
        event:
          type: string
        data:
          type: object
    HeartbeatEvent:
      allOf:
        - $ref: '#/components/schemas/StreamEventBase'
        - type: object
          properties:
            event:
              enum: [heartbeat]
            data:
              type: object
              required: [position]
              properties:
                position:
                  $ref: '#/components/schemas/Position'
    RecordStreamEvent:
      oneOf:
        - allOf:
            - $ref: '#/components/schemas/StreamEventBase'
            - type: object
              properties:
                event:
                  enum: [artifact]
                data:
                  $ref: '#/components/schemas/ArtifactSummary'
        - $ref: '#/components/schemas/HeartbeatEvent'
    SubmissionStreamEvent:
      oneOf:
        - allOf:
            - $ref: '#/components/schemas/StreamEventBase'
            - type: object
              properties:
                event:
                  enum:
                    - submission.received
                    - submission.accepted
                    - submission.rejected
                    - submission.skipped
                data:
                  $ref: '#/components/schemas/SubmissionResource'
        - $ref: '#/components/schemas/HeartbeatEvent'
...