# data-server/service/symposium_api/generated/routers/streams.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["streams"])


@router.get(
    "/{community}/streams/record",
    response_class=StreamingResponse,
    status_code=200,
    operation_id="streamRecord",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
        "429": {
            "model": Error,
        },
    },
)
async def stream_record(
    community: str,
    last__event__i_d: Optional[str] = Header(None, alias="Last-Event-ID"),
    caller: Caller = Depends(authorize("streamRecord")),
    service: Service = Depends(service_provider),
):
    """Follow the record as the gate accepts"""
    return StreamingResponse(
        sse_frames(
            service.stream_record(
                caller, community=community, last__event__i_d=last__event__i_d
            )
        ),
        media_type="text/event-stream",
        headers=SSE_HEADERS,
    )


@router.get(
    "/{community}/streams/submissions",
    response_class=StreamingResponse,
    status_code=200,
    operation_id="streamSubmissions",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
        "429": {
            "model": Error,
        },
    },
)
async def stream_submissions(
    community: str,
    last__event__i_d: Optional[str] = Header(None, alias="Last-Event-ID"),
    caller: Caller = Depends(authorize("streamSubmissions")),
    service: Service = Depends(service_provider),
):
    """Follow submissions and the gate's decisions"""
    return StreamingResponse(
        sse_frames(
            service.stream_submissions(
                caller, community=community, last__event__i_d=last__event__i_d
            )
        ),
        media_type="text/event-stream",
        headers=SSE_HEADERS,
    )